Renamed and deleted some reducers
